UtilMongo

+++ src/main/java/com/kristina/ecom/dao/ShoppingCartDAOMongo.java
@@ -3,44 +3,20 @@ package com.kristina.ecom.dao;
 import static com.mongodb.client.model.Filters.eq;
 
 import java.util.Date;
-import java.sql.ResultSet;
-import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.TimeZone;
 
 import org.bson.Document;
-import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
 
 import com.mongodb.MongoException;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
-import com.mongodb.client.model.Filters;
 import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.InsertOneResult;
-import com.mongodb.client.result.UpdateResult;
 import com.kristina.ecom.domain.Computer;
 import com.kristina.ecom.domain.ComputerBase;
 import com.kristina.ecom.domain.Product;
-
-      /*
-      {
-      "_id": ObjectId(""),
-      "user_id": string, --> prep for the User
-      "modifies": datetime,
-      "status": active, fulfilled, canceled,
-      "computers": [
-        "_id":  (this is for Computer, and the id is String. the string is manually created)
-        "_id": string, "description": string, "price": double
-        "products": [
-          "_id": (this is now the product id) integer, --> 
-          "quantity": integer
-         ]
-      ]
-      }
-      */ 
-
 public class ShoppingCartDAOMongo  implements DAO<String, Computer<String>> {
   private MongoDataSourceFactory dataSourceFactory;
   private MongoCollection<Document> collection;
@@ -63,8 +39,11 @@ public class ShoppingCartDAOMongo  implements DAO<String, Computer<String>> {
     return null;
   }
 
-  @Override
+  @Override // is this real all shopping carts or read all computers in  a specific shopping cart
   public List<Computer<String>> readAll() throws DAOException {
+    FindIterable<Document> shopDocuments = collection.find();
+    // List<Computer<String>> 
+
     return null;
   }
 
@@ -76,28 +55,30 @@ public class ShoppingCartDAOMongo  implements DAO<String, Computer<String>> {
       Computer<String> computer = toComputer(shoppingDocument);
       return computer;
     } else {
-      System.out.println("❌ Coudn't find the shopping cart with is: " + id);
+      System.out.println("❌ Coudn't find the shopping cart with id: " + id);
     }
 
     return null;
   }
 
   @Override
-  public int update(Computer<String> order) throws DAOException {
+  public int update(Computer<String> computer) throws DAOException {
     return 1;
   }
 
   @Override // delete the entire shopping cart
   public int delete(String id) throws DAOException {
     Document shopDocument = collection.find(eq("_id", new ObjectId(id))).first();
-
-    try {
-      DeleteResult result = collection.deleteOne(shopDocument);
-      System.out.println("❌ Deleted: " + result.getDeletedCount());
-      return (int) result.getDeletedCount();
-    } catch (MatchException ex) {
-      throw new DAOException("Coudn't delete the shopping cart", ex);
+    if (shopDocument != null) {
+      try {
+        DeleteResult result = collection.deleteOne(shopDocument);
+        System.out.println("❌ Deleted: " + result.getDeletedCount());
+        return (int) result.getDeletedCount();
+      } catch (MatchException ex) {
+        throw new DAOException("Coudn't delete the shopping cart", ex);
+      }
     }
+    return 0;
   }
 
   // from object to document
@@ -124,11 +105,11 @@ public class ShoppingCartDAOMongo  implements DAO<String, Computer<String>> {
   }
 
   private List<Document> toProductDocuments(List<Product<String>> products) {
-    ProductDAOMongo productDao = new ProductDAOMongo(); // using productDAO ??
+    UtilDAOMongo utilMongo = new UtilDAOMongo(); // use UtilMongo instead of calling the productDAO directly
     List<Document>  productDocuments = new ArrayList<>();
 
     for (Product<String> product : products) {
-      Document productDoc = productDao.toDocument(product);
+      Document productDoc = utilMongo.toDocument(product);
       
       productDocuments.add(productDoc);
     }
@@ -152,10 +133,10 @@ public class ShoppingCartDAOMongo  implements DAO<String, Computer<String>> {
 
   private List<Product<String>> toProducts(List<Document> productDocuments) {
     List<Product<String>> products = new ArrayList<>();
-    ProductDAOMongo productDao = new ProductDAOMongo();
+    UtilDAOMongo utilMongo = new UtilDAOMongo();
 
     for (Document productDoc : productDocuments) {
-      Product<String> product = productDao.toProduct(productDoc);
+      Product<String> product = utilMongo.toProduct(productDoc);
       products.add(product);
     }
     return products;
